+++ Demo_API/src/main/java/Dominio/Usuario.java
@@ -2,7 +2,7 @@ package Dominio;
 
 public class Usuario {
     private long id;
-    private int puntosConfianza;
+    private float puntosConfianza;
     private String gradoConfianza;
     private boolean activo;
 
@@ -10,7 +10,7 @@ public class Usuario {
         this.id = id;
     }
 
-    public void setPuntosConfianza(int puntosConfianza) {
+    public void setPuntosConfianza(float puntosConfianza) {
         this.puntosConfianza = puntosConfianza;
     }
